Assignment code refactored
