+++ src/Components/Assignment/QuizHeading.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const QuizHeading = (props) => {
+	return (
+		<div class = "headingContainer">
+	        <p> Course Name <br/>
+	            Quiz Name <br/>
+	            Assignment Info : Assignment No. 1 <br/>
+	            No. of Questions : 10 <br/>
+	            Attempts : 1 <br/>
+	            Maximum Marks : 20
+	        </p>
+	    </div>
+	);
+}
+
+
+export default QuizHeading;
